+++ app/mailers/user_mailer.rb
@@ -1,5 +1,7 @@
-class UserMailer < ApplicationMailer
+# frozen_string_literal: true
 
+# Sends out application emails to users.
+class UserMailer < ApplicationMailer
   def contact(name, email, message)
     subject = "#{name} - Website Message"
     @email_to = ENV['contact_email']
@@ -11,16 +13,15 @@ class UserMailer < ApplicationMailer
     @project_user = project_user
     @email_to = project_user.user.email
     mail(to: project_user.user.email,
-         subject: "#{project_user.creator.name} Allows You to #{project_user.editor? ? 'Edit' : 'View'} Project #{project_user.project.name}",
+         subject: "#{project_user.creator.name} Allows You to #{project_user.editor? ? 'Edit' : 'View'} #{project_user.project.name}",
          reply_to: project_user.creator.email)
   end
 
-  def invite_user_to_project(project_user)
+  def user_invited_to_project(project_user)
     @project_user = project_user
     @email_to = project_user.invite_email
     mail(to: project_user.invite_email,
-         subject: "#{project_user.creator.name} Invites You to #{project_user.editor? ? 'Edit' : 'View'} Project #{project_user.project.name}",
+         subject: "#{project_user.creator.name} Invites You to #{project_user.editor? ? 'Edit' : 'View'} #{project_user.project.name}",
          reply_to: project_user.creator.email)
   end
-
 end