+++ test/mailers/previews/user_mailer_preview.rb
@@ -1,18 +1,18 @@
-# Preview all emails at http://localhost:3000/rails/mailers/user_mailer
-class UserMailerPreview < ActionMailer::Preview
+# frozen_string_literal: true
 
+# Allows emails to be viewed at /rails/mailers
+class UserMailerPreview < ActionMailer::Preview
   def contact
-    UserMailer.contact("FirstName LastName", "example@test.com", "This is my message\nOn two lines!")
+    UserMailer.contact('FirstName LastName', 'example@test.com', "This is my message\nOn two lines!")
+  end
+
+  def user_invited_to_project
+    project_user = ProjectUser.where.not(invite_email: nil).first
+    UserMailer.user_invited_to_project(project_user)
   end
 
   def user_added_to_project
-    project_user = ProjectUser.where.not( invite_email: nil ).first
+    project_user = ProjectUser.where.not(user_id: nil).first
     UserMailer.user_added_to_project(project_user)
   end
-
-  def invite_user_to_project
-    project_user = ProjectUser.where.not( invite_email: nil ).first
-    UserMailer.invite_user_to_project(project_user)
-  end
-
 end