+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 ### Refactoring
 - Removed unused references to tumblr API
 - Cleaned up routes file
+- Cleaned up inviting and adding users to project
 
 ### Bug Fixes
 - Fixed a bug that prevented a project username from being reused if a deleted